requirements.txt
streamlit
PyPDF2
google-generativeai
python-dotenv
sentence-transformers
torch
langchain
langchain-core
langchain-community
langchain-google-genai
